LPS-53386 Undo all the rule changes from previous Arquillian integration
