+++ portal-test-internal/src/com/liferay/portal/test/rule/LogAssertionHandler.java
@@ -15,8 +15,6 @@
 package com.liferay.portal.test.rule;
 
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.test.rule.executor.LogAssertionExecutor;
-import com.liferay.portal.test.rule.executor.LogAssertionExecutorUtil;
 
 import java.util.logging.Handler;
 import java.util.logging.Level;
@@ -52,13 +50,8 @@ public class LogAssertionHandler extends Handler {
 			sb.append(", message=");
 			sb.append(logRecord.getMessage());
 
-			LogAssertionExecutor logAssertionExecutor =
-				LogAssertionExecutorUtil.getInstance();
-
-			logAssertionExecutor.caughtFailure(
-				Thread.currentThread(),
-				new AssertionError(
-					new AssertionError(sb.toString(), logRecord.getThrown())));
+			LogAssertionTestRule.caughtFailure(
+				new AssertionError(sb.toString(), logRecord.getThrown()));
 		}
 	}
