--- portal-test-internal/src/com/liferay/portal/test/rule/executor/InitTestLiferayContextExecutorImpl.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.test.rule.executor;
-
-import com.liferay.portal.kernel.test.rule.executor.InitTestLiferayContextExecutor;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.util.InitUtil;
-import com.liferay.portal.util.PropsUtil;
-
-import java.util.List;
-
-/**
- * @author Cristina González
- */
-public class InitTestLiferayContextExecutorImpl
-	implements InitTestLiferayContextExecutor {
-
-	@Override
-	public void init() {
-		if (InitUtil.isInitialized()) {
-			return;
-		}
-
-		System.setProperty("catalina.base", ".");
-
-		List<String> configLocations = ListUtil.fromArray(
-			PropsUtil.getArray(PropsKeys.SPRING_CONFIGS));
-
-		InitUtil.initWithSpring(configLocations, true);
-
-		if (System.getProperty("external-properties") == null) {
-			System.setProperty("external-properties", "portal-test.properties");
-		}
-	}
-
-}
