--- portal-test-internal/src/com/liferay/portal/test/rule/executor/UniqueStringRandomizerBumperExecutorImpl.java
@@ -1,31 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.test.rule.executor;
-
-import com.liferay.portal.kernel.test.randomizerbumpers.UniqueStringRandomizerBumper;
-import com.liferay.portal.kernel.test.rule.executor.UniqueStringRandomizerBumperExecutor;
-
-/**
- * @author Cristina González
- */
-public class UniqueStringRandomizerBumperExecutorImpl
-	implements UniqueStringRandomizerBumperExecutor {
-
-	@Override
-	public void reset() {
-		UniqueStringRandomizerBumper.reset();
-	}
-
-}
